Various
- impl `LineLimit`
- Move more general styles into interface/style.
