+++ src/interface/bin/mod.rs
@@ -1,4 +1,3 @@
-pub mod color;
 pub mod style;
 mod text_state;
 
@@ -2291,10 +2290,22 @@ impl Bin {
             back_color.a *= opacity;
         }
 
-        let border_radius_tl = style.border_radius_tl.px_width([width, height]).unwrap_or(0.0);
-        let border_radius_tr = style.border_radius_tr.px_width([width, height]).unwrap_or(0.0);
-        let border_radius_bl = style.border_radius_bl.px_width([width, height]).unwrap_or(0.0);
-        let border_radius_br = style.border_radius_br.px_width([width, height]).unwrap_or(0.0);
+        let border_radius_tl = style
+            .border_radius_tl
+            .px_width([width, height])
+            .unwrap_or(0.0);
+        let border_radius_tr = style
+            .border_radius_tr
+            .px_width([width, height])
+            .unwrap_or(0.0);
+        let border_radius_bl = style
+            .border_radius_bl
+            .px_width([width, height])
+            .unwrap_or(0.0);
+        let border_radius_br = style
+            .border_radius_br
+            .px_width([width, height])
+            .unwrap_or(0.0);
         let max_radius_t = border_radius_tl.max(border_radius_tr);
         let max_radius_b = border_radius_bl.max(border_radius_br);
         let max_radius_l = border_radius_tl.max(border_radius_bl);
@@ -2892,7 +2903,7 @@ fn z_unorm(z: i16) -> f32 {
 }
 
 #[inline(always)]
-fn lerp(t: f32, a: f32, b: f32) -> f32 {
+pub(crate) fn lerp(t: f32, a: f32, b: f32) -> f32 {
     (t * b) + ((1.0 - t) * a)
 }
